+++ visualizer/src/Controls/ImageControls/RawControls/ColorPalette.js
