+++ visualizer/src/Controls/ImageControls/RawControls/GrayscaleControls.js
