+++ visualizer/src/Controls/ImageControls/RawControls/RawControls.js
