+++ visualizer/src/Controls/Toolbar/UndoRedo.js
@@ -6,7 +6,7 @@ import UndoIcon from '@material-ui/icons/Undo';
 import { useSelector } from '@xstate/react';
 import React from 'react';
 
-import { useUndo } from '../ServiceContext';
+import { useUndo } from '../../ServiceContext';
 
 const useStyles = makeStyles({
   button: {
