+++ visualizer/src/Instructions/SelectInstructions.js
@@ -1,7 +1,9 @@
 import Typography from '@material-ui/core/Typography';
 import React from 'react';
 
-import SelectedPalette, { SwapButton } from '../ControlPanel/SelectedPalette';
+import SelectedPalette, {
+  SwapButton,
+} from '../Controls/Toolbar/SelectedPalette';
 
 function SelectInstructions() {
   return (
